+++ tools/src/native_app/app_settings.rs
@@ -2,5 +2,5 @@ use diligent::RenderDeviceType;
 
 pub trait AppSettings {
     fn get_render_device_type(&self) -> RenderDeviceType;
-    fn get_window_dimensions(&self) -> (u16, u16);
+    fn get_window_dimensions(&self) -> (u32, u32);
 }
